Front end design update and refactoring
