+++ src/App.js
@@ -1,22 +1,35 @@
 import "./App.css";
-import { React } from "react";
-import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
+import { React, useState, Component } from "react";
+import { BrowserRouter as Router, Switch, Route, Link, Redirect } from "react-router-dom";
 import Home from "./Components/Home";
 import About from "./Components/About";
-import Navbar from "./Components/NavBar";
+import LoginForm from "./Components/Login";
 
 function App() {
+  const getQueryParam = (paramName) => {
+    const queryString = window.location.search;
+    const urlParams = new URLSearchParams(queryString);
+    return urlParams.get(paramName);
+  };
 
+  const [userId, setUserId] = useState(getQueryParam("user_id"));
+  const [username, setUsername] = useState(getQueryParam("username"));
+
+  
   return (
     <Router>
       <Switch>
+        
         <Route exact path="/">
-          <Home />
+          {!userId ? <LoginForm /> : <Redirect to='/home'/>}
+          
+        </Route>
+        <Route path='/home'>
+          <Home  userId={userId} username={username}/>
         </Route>
         <Route exact path="/about">
           <About />
         </Route>
-        <Route path="/auth"></Route>
       </Switch>
     </Router>
   );
